GetValidatedArgument[Async] uses context.RequestServices
Using the schema to access the root service provider made it impossible to use validators with a scoped lifetime, or with scoped dependencies.  By using context.RequestServices we can access the service provider that's appropriate for the current scope (Asp.Net request scope, or a child scope created for a field resolver).